Restructured the routing so it's easier to understand and less bulky.
