+++ back-end/server.js
@@ -26,148 +26,17 @@ const express = require('express');
 require('dotenv').config();
 const API_PORT = process.env.PORT || 3001;
 const app = express();
-const router = express.Router();
-
-const database = require('./database.js');
 app.use(bodyParser.json());
 
-// ============================================================================
-// '/api/image' routes
+const MainRouter = express.Router();
 
-router.get('/image/:imageId/view', (req, res) => {
-	// Get a specific image's direct url
-	console.log("GET /image/:imageID/view");
-	console.log(req.params);
-	// Might be a good idea to change this to getPhotoReference later on
-	database.getPhotoData('test_user', req.params.imageId).then((image) => {
-		console.log(image.reference);
-		res.send(image.reference);
-	})
-	// res.sendStatus(200);
-});
-
-router.get('/image/:imageId', (req, res) => {
-	// Get an image - Reference, Name, DateTime, Description
-	console.log("GET /image/:imageID");
-	console.log(req.params);
-	database.getPhotoData('test_user', req.params.imageId).then((image) => {
-		console.log(image);
-		res.send(image);
-	});
-	// res.send();
-	// res.sendStatus(200);
-	// res.end();
-});
-
-router.post('/image/', (req, res) =>{
-	// Add a new image - Likely will need upload or some kind of url
-	console.log("POST /image/");
-	console.log(req.params);
-	console.log(req.query);
-	res.sendStatus(200);
-});
-
-router.delete('/image/', (req, res) => {
-	// Delete an image entry. If it has an uploaded image deal with that.
-	console.log("DELETE /image/");
-	console.log(req.params);
-	res.sendStatus(200);
-});
-
-// ============================================================================
-// '/api/album' routes 
-
-// '/api/album/:albumId'
-router.get   ('/album/:albumId?', (req, res) => {
-	// Get an album with a specified ID
-	console.log("GET /album/:albumID");
-	console.log(req.params);
-
-	// get 'un'-album if no id specified
-	if (!req.params.albumId) albumId = 'un';
-	// console.log(req.query);
-	database.getAlbumData('test_user', 'un').then((album) => {
-		// TODO: Work out what to send here?
-
-		console.log(album);
-		console.log(album.photos[0]._key);
-		console.log(album.photos[0].exists);
-
-		/*
-		album.photos[0].get().then(doc => {
-			console.log(doc);
-			if (!doc.exists) {
-				console.log('No such document!');
-			} else {
-				console.log('Document data:', doc.data());
-			}
-		}).catch(err => {
-			console.log('Error getting document', err);
-		});
-		*/
-	});
-	res.sendStatus(200);
-});
-router.post  ('/album/', (req, res) => {
-	// Create a new album, and return its ID
-	console.log("POST /album/:albumID");
-	console.log(req.params);
-	console.log(req.query);
-	res.sendStatus(200);
-});
-router.put   ('/album/:albumId', (req, res) => {
-	// Update an album (May be removed in preference to PATCH)
-	console.log("PUT /album/:albumID");
-	console.log(req.params);
-	res.sendStatus(200);
-});
-router.patch ('/album/:albumId', (req, res) => {
-	// Update an album
-	console.log("PATCH /album/:albumID");
-	console.log(req.params);
-	res.sendStatus(200);
-});
-router.delete('/album/:albumId', (req, res) => {
-	// Delete an album
-	console.log("DELETE /album/:albumID");
-	console.log(req.params);
-	res.sendStatus(200);
-});
-
-// '/api/album/:albumId/:pageId'
-router.get   ('/album/:albumId/:pageId', (req, res) => {
-	// Get a page of an album
-	console.log("GET /album/:albumID/:pageId");
-	console.log(req.params);
-	res.sendStatus(200);
-});
-router.post  ('/album/:albumId/:pageId', (req, res) => {
-	// Create a new album page
-	console.log("POST /album/:albumID/:pageId");
-	console.log(req.params);
-	res.sendStatus(200);
-});
-router.put   ('/album/:albumId/:pageId', (req, res) => {
-	// Update an album page (might be useful eg. for copy/paste)
-	console.log("PUT /album/:albumID/:pageId");
-	console.log(req.params);
-	res.sendStatus(200);
-});
-router.patch ('/album/:albumId/:pageId', (req, res) => {
-	// Update an album page (just a small part)
-	console.log("PATCH /album/:albumID/:pageId");
-	console.log(req.params);
-	res.sendStatus(200);
-});
-router.delete('/album/:albumId/:pageId', (req, res) => {
-	// Delete an album page
-	console.log("DELETE /album/:albumID/:pageId");
-	console.log(req.params);
-	res.sendStatus(200);
-});
+// Route Handlers
+MainRouter.use('/image', require('./routes/image').router);
+MainRouter.use('/album', require('./routes/album').router);
 
 // prepend '/api' for our http requests 
-app.use('/api', router);
+app.use('/api', MainRouter);
+
 
 // launch our backend into a port
 app.listen(API_PORT, () => console.log('LISTENING ON PORT '+API_PORT));
